f2

+++ controllers/book.controller.js
@@ -1,11 +1,13 @@
 const  Books  = require('../models/books.model');
-const category = require('../models/category.model');
-
+const Category = require('../models/category.model');
+const mongoose = require('mongoose');
 ///////////////////////////////////////////////////
 //admin
 //user as a admin  addBook,allBook, editBook,showSingleBookById,DeleteBook
 addBook = async (req, res) => {
     try {
+      const category = await Category.findById(req.body.category);
+      if(!category) return res.status(400).send('Invalid Category')
       const bookData = new Books(req.body)
       await bookData.save()
       res.status(200).send({
@@ -42,19 +44,29 @@ addBook = async (req, res) => {
   editBook = async (req, res) => {
     try {
       const _id = req.params.id
+      if(!_id){
+        return res.status(400).send('Invalid book Id')
+      }
+      const category = await Category.findById(req.body.category);
+      if(!category) return res.status(400).send('Invalid Category')
+      
       const updatedData = req.body;
       await Books.findByIdAndUpdate(_id,
         {
           authors: updatedData.authors,
           price: updatedData.price,
           bookName: updatedData.bookName,
-          type: updatedData.type,
           cover: updatedData.cover,
           desc: updatedData.desc,
           numPage: updatedData.numPage,
           countInStock: updatedData.countInStock,
           numReviews: updatedData.numReviews,
           language: updatedData.language,
+          category: updatedData.category,
+          richDes: updatedData.richDes,
+          dateInsert:updatedData.dateInsert,
+          rating: updatedData.rating,
+          isFeatured: updatedData.isFeatured,
         }, { new: true },);
         res.status(200).send({
           status: true,
@@ -93,6 +105,9 @@ addBook = async (req, res) => {
   
   DeleteBook = async (req, res) => {
     try {
+      if(!mongoose.isValidObjectId(req.params.id)) return res.status(400).send('invalid book id')
+      const category = await Category.findById(req.body.category);
+      if(!category) return res.status(400).send('Invalid Category')
       const _id = req.params.id
       const deleteBook = await Books.findByIdAndDelete(_id)
       if (!deleteBook) res.send('book not found')
